+++ tests/EventListener/CommandListenerTest.php
@@ -40,7 +40,7 @@ class CommandListenerTest extends TestCase
     /**
      * @inheritDoc
      */
-    protected function setUp()
+    protected function setUp(): void
     {
         $this->api = $this->createMock(BotApi::class);
         $this->commandRegistry = $this->createMock(CommandRegistry::class);