+++ src/Event/UpdateEvent.php
@@ -9,9 +9,9 @@
  * file that was distributed with this source code.
  */
 
-namespace BoShurik\TelegramBotBundle\Event\Telegram;
+namespace BoShurik\TelegramBotBundle\Event;
 
-use Symfony\Component\EventDispatcher\Event;
+use Symfony\Contracts\EventDispatcher\Event;
 use TelegramBot\Api\Types\Update;
 
 class UpdateEvent extends Event