+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 CHANGELOG
 =========
 
+4.0.0 (2019-11-21)
+------------------
+
+* Drop support of `symfony/symfony` < 4.4
+* Move `\BoShurik\TelegramBotBundle\Event\Telegram\UpdateEvent` to `\BoShurik\TelegramBotBundle\Event\UpdateEvent`
+* Move `\BoShurik\TelegramBotBundle\Event\TelegramEvents` to `\BoShurik\TelegramBotBundle\Event\WebhookEvent`
+* Removed `\BoShurik\TelegramBotBundle\Event\TelegramEvents`
+
 3.1.0 (2019-10-16)
 ------------------
 
@@ -14,7 +22,7 @@ CHANGELOG
 
 * Command system now works with `Update` object instead of `Message` (#14)
 * Drop support of php5
-* Drop support of symfony/symfony < 3.4
+* Drop support of `symfony/symfony` < 3.4
 * Change bundle alias from `bo_shurik_telegram_bot` to `boshurik_telegram_bot`
 * Split `bin/console telegram:webhook` command to `bin/console telegram:webhook:set` 
 and `bin/console telegram:webhook:unset`